# server/docs/swagger.yaml
basePath: /v1
definitions:
...
  controller.CreateReviewRequest:
    properties:
      film_id:
        type: integer
      rating:
        maximum: 100
        minimum: 0
        type: integer
      review_text:
        type: string
      user_id:
        type: integer
    required:
    - film_id
    - rating
    - review_text
    - user_id
    type: object
...
  controller.UpdateReviewRequest:
    properties:
      film_id:
        type: integer
      rating:
        maximum: 100
        minimum: 0
        type: integer
      review_id:
        type: integer
      review_text:
        type: string
      user_id:
        type: integer
    required:
    - film_id
    - rating
    - review_id
    - review_text
    - user_id
    type: object
...
  response.Response:
    description: Structure for a standard API response
    properties:
      data: {}
      error:
        example: any error
        type: string
      status:
        example: success/error
        type: string
    type: object
info:
  contact:
    name: Evdokimov Igor
    url: https://t.me/epelptic
  description: API for potatorate site
  title: Film-catalog API
  version: 1.0.0
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Создает новый отзыв с указанными данными
      parameters:
      - description: Данные отзыва
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Создание нового отзыва
      tags:
      - review
    put:
      consumes:
      - application/json
      description: Обновляет данные отзыва
      parameters:
      - description: Данные отзыва
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Обновление отзыва
      tags:
      - review
  /reviews/{id}:
    delete:
      description: Удаляет отзыв по указанному ID
      parameters:
      - description: ID отзыва
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Удаление отзыва
      tags:
      - review
    get:
      description: Возвращает информацию о отзыве по указанному ID
      parameters:
      - description: ID отзыва
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получение отзыва по ID
      tags:
      - review
  /reviews/film/{film_id}:
    get:
      description: Возвращает список отзывов для указанного фильма
      parameters:
      - description: ID фильма
        in: query
        name: film_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получение отзывов по ID фильма
      tags:
      - review
  /reviews/reviewer/{reviewer_id}:
    get:
      description: Возвращает список отзывов, оставленных указанным пользователем
      parameters:
      - description: ID пользователя
        in: query
        name: reviewer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Получение отзывов по ID пользователя
      tags:
      - review
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"